SetColor now works

+++ src/Color.h
@@ -1,13 +1,15 @@
 #ifndef COLOR_H_
 #define COLOR_H_
 
+#include <SDL.h>
+
 class Color
 {
 public:
-    int red;
-    int blue;
-    int green;
-    int alpha;
+    Uint8 red;
+    Uint8 blue;
+    Uint8 green;
+    Uint8 alpha;
 
     Color() : red(255), blue(255), green(255), alpha(255) {}
 };
